requirements.txt
click==8.1.8
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
typer==0.15.4
typing_extensions==4.13.2
